feature(Order, Inventory, Payment): - 新增订单模块，完成基础订单功能 - 新增库存模块，取代原本Product中的stock字段 - 新增支付模块，支持支付宝支付

+++ backend/src/main/java/cn/edu/nju/TomatoMall/models/dto/product/ProductCreateRequest.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 import lombok.NonNull;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -15,7 +16,7 @@ public class ProductCreateRequest {
     String title;
     String description;
     List<MultipartFile> images;
-    double price;
+    BigDecimal price;
     int stock;
     int storeId;
     Map<String, String> specifications;
